feat: added actions & reducers delete template
