+++ src/services/reducers/index.ts
@@ -13,8 +13,7 @@ import { getBotsReducer } from './bots/getBots';
 import { addBotReducer } from './bots/addBot';
 
 import { getTemplatesBotsReducer } from './bots/getTemplatesBots';
-import { addTemplatesBotReducer } from './bots/addTemplatesBot';
-import { deleteTemplatesBotReducer } from './bots/deleteTemplatesBot';
+import { addTemplatesBotReducer } from './bots/templatesBot';
 
 import { getPlatformsReducer } from './platforms/getPlatforms';
 import { socketReducer } from './socket/socketReducer';
@@ -31,7 +30,6 @@ const rootReducer = combineReducers({
   addBot: addBotReducer,
   getTemplatesBots: getTemplatesBotsReducer,
   addTemplatesBot: addTemplatesBotReducer,
-  deleteTemplatesBot: deleteTemplatesBotReducer,
   getPlatforms: getPlatformsReducer,
   websocket: socketReducer,
   toggleMessengersPopup: toggleMesPopup,
