+++ src/services/types/index.ts
@@ -20,8 +20,8 @@ import { TGetBotsActions } from '../actions/bots/getBot';
 import { TAddBotState } from '../reducers/bots/addBot';
 import { TAddBotActions } from '../actions/bots/addBot';
 
-import { TAddTemplatesBotState } from '../reducers/bots/addTemplatesBot';
-import { TAddTemplatesBotActions } from '../actions/bots/addTemplatesBot';
+import { TAddTemplatesBotState } from '../reducers/bots/templatesBot';
+import { TAddTemplatesBotActions } from '../actions/bots/templatesBot';
 
 import { TDeleteTemplatesBotState } from '../reducers/bots/deleteTemplatesBot';
 import { TDeleteTemplatesBotActions } from '../actions/bots/deleteTemplatesBot';
